pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "qrcode>=8.0",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.43.1",
]
